+++ internal/test/e2e/autosdk/main.go
@@ -1,6 +1,8 @@
 // Copyright The OpenTelemetry Authors
 // SPDX-License-Identifier: Apache-2.0
 
+// Package autosdk is a testing application for the
+// [go.opentelemetry.io/auto/sdk] package.
 package main
 
 import (